move CommandMenuContextRecordChipAvatars in another file
